refactor: 拆分 HubAssistantCard 和 AssistantDetailModal 为独立组件
